fix: amount 타입 balance와 맞춰서 Long으로 통일

+++ payment-service/src/main/java/com/lgcns/bebee/payment/application/client/TossPaymentsClient.java
@@ -1,7 +1,7 @@
 package com.lgcns.bebee.payment.application.client;
 
 public interface TossPaymentsClient {
-    TossPaymentResponse confirmPayment(String paymentKey, String orderId, Integer amount);
+    TossPaymentResponse confirmPayment(String paymentKey, String orderId, Long amount);
 
 //    TossPaymentResponse getPayment(String paymentKey);
 //
